Bug: runtime: var _stp_allocated_net_memory would overflow easily
Changed its type from int to long. We saw it overflow in a VMWare guest
which has 128 CPUs configured.

The patch was originally drafted by Junlong Li.

+++ runtime/linux/alloc.c
@@ -14,7 +14,7 @@
 #include <linux/percpu.h>
 #include "stp_helper_lock.h"
 
-static int _stp_allocated_net_memory = 0;
+static long _stp_allocated_net_memory = 0;
 /* Default, and should be "safe" from anywhere. */
 #define STP_ALLOC_FLAGS ((GFP_KERNEL | __GFP_NORETRY | __GFP_NOWARN) \
 			 & ~__GFP_WAIT)
@@ -45,7 +45,7 @@ static int _stp_allocated_net_memory = 0;
  * would be nice, but DEBUG_MEM is only for testing.
  */
 
-static int _stp_allocated_memory = 0;
+static long _stp_allocated_memory = 0;
 
 #ifdef DEBUG_MEM
 static STP_DEFINE_SPINLOCK(_stp_mem_lock);
